ManageCourses and ManageLecturers Controller Sessions all updated

+++ src/main/java/sg/edu/iss/ca/controller/AdminManageLecturersController.java
@@ -1,5 +1,7 @@
 package sg.edu.iss.ca.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,28 +22,45 @@ public class AdminManageLecturersController {
 	AdminLecturerInterface alService;
 	
 	@RequestMapping("/list")
-	public String list(Model model) {
+	public String list(Model model, HttpSession session) {
+		if (session.getAttribute("asession") == null) {
+			return "forward:/admin/login";
+		}
+		
 		model.addAttribute("lecturers", alService.findALlLecturers());
 		return "admin/adminLecturer";
 	}
 	
 	@PostMapping("/addLecturer")
-	public String addUser(Model model) 
+	public String addUser(Model model, HttpSession session) 
 	{
+		if (session.getAttribute("asession") == null) {
+			return "forward:/admin/login";
+		}
+		
 		Lecturer newLecturer = new Lecturer();
 		model.addAttribute("newLecturer", newLecturer);
 		return "admin/adminLecturerAdd";
 	}
 	
 	@PostMapping("/saveLecturer")
-	public String saveUser(@ModelAttribute("newLecturer") Lecturer newLecturer)
+	public String saveUser(@ModelAttribute("newLecturer") Lecturer newLecturer, HttpSession session)
 	{
+		if (session.getAttribute("asession") == null) 
+		{
+			return "forward:/admin/login";
+		}
+		
 		alService.createLecturerRecord(newLecturer);
 		return "forward:/manageLecturers/list";
 	}
 	
 	@RequestMapping("/updateLecturer")
-	public String editForm(@RequestParam("lecturerId") int lecturerId, Model model) {
+	public String editForm(@RequestParam("lecturerId") int lecturerId, Model model, HttpSession session) {
+		if (session.getAttribute("asession") == null) {
+			return "forward:/admin/login";
+		}
+		
         Lecturer updateLecturer = alService.findLecturerById(lecturerId);
         model.addAttribute("newLecturer", updateLecturer);
         model.addAttribute("update", true);
@@ -49,7 +68,10 @@ public class AdminManageLecturersController {
 	}
 	
     @GetMapping("/deleteLecturer")
-    public String delete(@RequestParam("lecturerId") int lecturerId){
+    public String delete(@RequestParam("lecturerId") int lecturerId, HttpSession session){
+		if (session.getAttribute("asession") == null) {
+			return "forward:/admin/login";
+		}
     	alService.deleteLecturer(lecturerId);
         return "redirect:/manageLecturers/list";
     }
